requirements.txt
elasticsearch==8.9.0
Flask==2.3.2
urllib3==2.0.4
Jinja2==3.1.2
Werkzeug==2.3.6
pytest==7.3.1
python-dotenv==1.0.1

